Fix: update task status and priority routes

+++ index.js
@@ -74,30 +74,46 @@ app.post("/tasks", authMiddleware, async (req, res) => {
 });
 
 // delete task request
-app.delete("/task/:id", authMiddleware, async (req, res) => {
-  await Task.findOneAndDelete({ _id: req.params.id.userId });
-  res.json({ message: "Task deleted" });
+app.delete("/tasks/:id", authMiddleware, async (req, res) => {
+  try {
+    await Task.findOneAndDelete({ _id: req.params.id, userId: req.userId });
+    res.json({ message: "Task deleted" });
+  } catch (e) {
+    res.status(500).json({ message: "Server error" });
+  }
 });
 
+
 // update task request
-app.patch("/tasks", authMiddleware, async (req, res) => {
+app.patch("/tasks/:id", authMiddleware, async (req, res) => {
   const { status } = req.body;
+  try{
   const task = await Task.findOneAndUpdate(
     { _id: req.params.id, userId: req.userId },
     { status },
     { new: true }
+  
   );
   if (!task) return res.status(404).json({ message: "Task not found" });
+  res.json(task);
+  } catch (e) {
+    res.status(500).json({ message: "Server error" });
+  }
 });
 
 app.patch("/tasks/:id/priority", authMiddleware, async (req, res) => {
   const { priority } = req.body;
+  try {
   const task = await Task.findOneAndUpdate(
     { _id: req.params.id, userId: userId },
     { priority },
     { new: true }
   );
   if (!task) return res.status(404).json({ message: "Task not found" });
+  res.json(task);
+} catch (e) {
+  res.status(500).json({ message: "Server error" });
+}
 });
 
 app.listen(PORT, () => {
